+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Ask server for a single result when searching for an account at subscription page.
+
 ## [4.8.0] - 2025-08-18
 
 ### Added