+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Removed extra space between post preview and attachments.
+- Fixed image overflow in lightbox.
 
 ## [4.8.0] - 2025-08-18